fix Citizen class

+++ src/main/java/ru/plidia/stateapp/entity/Citizen.java
@@ -2,12 +2,13 @@ package ru.plidia.stateapp.entity;
 
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import static ru.plidia.stateapp.service.DataGeneration.generateRandomAge;
 import static ru.plidia.stateapp.service.DataGeneration.generateRandomName;
 
 public class Citizen {
-    static AtomicInteger nextId = new AtomicInteger();
+    static AtomicLong nextId = new AtomicLong();
     private Long id;
     private String name;
     private String surname;
@@ -17,7 +18,7 @@ public class Citizen {
     }
 
     public Citizen(String name, String surname, int age) {
-        this.id = (long) nextId.incrementAndGet();
+        this.id = nextId.incrementAndGet();
         this.name = generateRandomName(name);
         this.surname = generateRandomName(surname);
         this.age = generateRandomAge(age);
